refact(services): Organize folder structure
